Move GoodInstance to trading
