+++ src/openvic-simulation/InstanceManager.hpp
@@ -5,7 +5,7 @@
 #include "openvic-simulation/console/ConsoleInstance.hpp"
 #include "openvic-simulation/country/CountryInstance.hpp"
 #include "openvic-simulation/diplomacy/CountryRelation.hpp"
-#include "openvic-simulation/economy/GoodInstance.hpp"
+#include "openvic-simulation/economy/trading/GoodInstance.hpp"
 #include "openvic-simulation/economy/trading/MarketInstance.hpp"
 #include "openvic-simulation/map/MapInstance.hpp"
 #include "openvic-simulation/map/Mapmode.hpp"
