+++ src/openvic-simulation/economy/trading/GoodInstance.cpp
