+++ src/openvic-simulation/economy/trading/GoodInstance.hpp
